+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- new rule `require-sri`.
 - add `<slot>` element.
 
 ## 0.24.1 (2019-03-26)
@@ -14,8 +15,6 @@
 - fix broken import causing typescript API users getting build errors.
 
 ## 0.24.0 (2019-03-26)
-
-- new rule `require-sri`.
 
 ### Features